requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
PyYAML>=6.0
ultralytics>=8.0.0  # YOLOv8
google-generativeai>=0.4.0
requests>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
